% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{load.RData}
\alias{load.RData}
\title{Helper function that loads a specific file and returns the variables therein}
\usage{
load.RData(filename)
}
\arguments{
\item{filename}{The path to the RData file}
}
\value{
Returns the content of the RData file as data frame
}
\description{
Helper function that loads a specific file and returns the variables therein
}

