% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.tab.R
\name{read.tab}
\alias{read.tab}
\title{Read tab}
\usage{
read.tab(filename, outdir = ".", RData = FALSE, tag = "tab2csv")
}
\arguments{
\item{filename}{The filename of the STATA log file}

\item{outdir}{Directory for output}

\item{RData}{store output in RData}

\item{tag}{specifies what tag to look for in the STATA log file}
}
\description{
Reads STATA tables created by \code{tab x y} and \code{tab x, sum(y)}. For
each \code{tab} command found, this function will create a CSV or RData file.
For this function to work there must be a statement
\code{di "{tag} filename.csv"} preceding every \code{tab} command.
}

