% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readLog.R
\name{read.tabstat}
\alias{read.tabstat}
\title{Read tabstat}
\usage{
read.tabstat(filename, outdir = ".", RData = FALSE)
}
\arguments{
\item{filename}{The filename of the STATA log file}

\item{outdir}{Directory for output}

\item{RData}{store output in RData}
}
\description{
Reads STATA tables created by \code{tabstat y, by(x)}. For each \code{tab}
command found, this function will create a CSV or RData file. For this function to
work there must be a statement \code{di "tabstat2csv filename.csv"}
preceding every \code{tabstat} command.
}

