% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{trim}
\alias{trim}
\title{Trim whitespaces from a string}
\usage{
trim(x)
}
\arguments{
\item{x}{The string to be stripped}
}
\value{
A string without whitespaces at the beginning and end
}
\description{
This function removes whitespaces from the beginning and end of a string
}

