% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readLog.R
\name{read.reg}
\alias{read.reg}
\title{Read regression output}
\usage{
read.reg(filename, outdir = ".", RData = TRUE)
}
\arguments{
\item{filename}{The filename of the STATA log file}

\item{outdir}{Directory for output}

\item{RData}{store output in RData}
}
\description{
Reads STATA tables created by linear regressions. For
each \code{(a)reg} command found, this function will create an \code{.RData}
file that contains the list \code{tab.reg} with the regression output.
}

