% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readLog.R
\name{read.tab}
\alias{read.tab}
\title{Read tab}
\usage{
read.tab(filename, outdir = ".")
}
\arguments{
\item{filename}{The filename of the STATA log file}

\item{outdir}{Directory for CSV output}
}
\description{
Reads STATA tables created by \code{tab x y`} and \code{tab x, sum(y)`}. For
each \code{tab} command found, this function will create a CSV file. For this
function to work there must be a statement \code{"di tab2csv filename.csv"}
preceding every \code{tab} command.
}

